# lab08/swagger.yaml
swagger: "2.0"
info:
  description: "lab08t04"
  version: "1.0.0"
  title: "Contact list"
host: "localhost:8080"
basePath: "/lab08/hw8"
tags:
- name: "eshop"
  description: "Yoghurt eshop"
  externalDocs:
    description: "Great yoghurt eshop"
    url: "http://localhost:8080"
schemes:
- "http"
paths:
  /eshop/total-amount/{id}:
    get:
      tags:
      - "eshop"
      summary: "Shows total amount for order with id from path"
      operationId: "totalAmount"
      parameters:
      - name: id
        type: number
        in: path
        required: true
      produces:
      - "text"
      responses:
        "200":
          description: "All ok"
          
  /eshop/shopping-cart/{id}:
    get:
      tags:
      - "eshop"
      summary: "Show all yoghurts for order with id from path"
      operationId: "showCart"
      produces:
      - "application/xml"
      - "application/json"
      parameters:
      - name: id
        type: number
        in: path
        required: true
      responses:
        "200":
          description: "Cart shown"
    delete:
      tags:
      - "eshop"
      summary: "Deletes order with id from path"
      operationId: "deleteOrder"
      produces:
      - "text"
      parameters:
      - name: id
        type: number
        in: path
        required: true
      responses:
        "200":
          description: "Order deleted"
      
    post:
      tags:
      - "eshop"
      operationId: "addSomeYoghs"
      summary: "Add some yoghurts of a single flavor to shopping cart."
      parameters:
      - name: id
        type: number
        in: path
        required: true
      - in: body
        name: yoghurtOrder
        description: "Specify the number and the flavor of yoghurts you want to add to cart."
        schema: 
          type: object
          properties:
            count:
              type: number
            flavorID:
              type: number
          xml:
            name: yoghurtOrder
      consumes:
      - "application/xml"
      - "application/json"
      responses:
        "200":
          description: "Yoghurts added to cart"
        "400":
          description: "Invalid order id"
        "404":
          description: "Unknown order id, create new order by calling /eshop/customer"
      
          
  /eshop/customer:
    put:
      tags:
      - "eshop"
      operationId: "newCustomer"
      consumes:
      - "application/xml"
      - "application/json"
      summary: "After sending customer credentials returns the id of an order that is used while shopping."
      produces:
      - "text"
      parameters:
      - in: body
        name: customer
        description: "address and name of new customer"
        schema: 
          type: object
          properties:
            name:
              type: string
            address:
              type: string
          xml:
            name: customer
      responses:
        "201":
          description: "Customer added"
      
          
  # /eshop/{count}:
  #   put:
  #     tags:
  #     - "eshop"
  #     operationId: "addSomeYoghs"
  #     parameters:
  #     - name: count
  #       type: number
  #       in: path
  #       required: true
  #     - in: body
  #       name: yoghurtOrder
  #       description: "dlflhdsklshf"
  #       schema: 
  #         type: object
  #         properties:
  #           count:
  #             type: number
  #           flavorID:
  #             type: number
  #         xml:
  #           name: yoghurtOrder
        
            
          